updated code, fixed guncon, fixed reset loop bug
Found the issue with the guncon, it was a missing break in the switch...
Also found a reset loop bug when you disconnect the controller from the PS1 after resetting it. The old data was never cleared so the arduino thought it receives a reset command and resets the PS1.